chore: rename old e2e tests for consistency
