+++ ui-tests/__test__/flows/navigate-prompts/navigate-prompts-up.ts
