+++ ui-tests/__test__/main.spec.ts
@@ -20,9 +20,9 @@ import { renderInformationCard } from './flows/render-information-card';
 import { renderTabbedCard } from './flows/render-tabbed-card';
 import { welcomeMode } from './flows/welcome-mode';
 import { renderButtons } from './flows/render-buttons';
-import { navigatePromptsDown } from './flows/navigate-prompts/navigatePromptsDown';
-import { navigatePromptsUp } from './flows/navigate-prompts/navigatePromptsUp';
-import { navigatePromptsToEmpty } from './flows/navigate-prompts/navigatePromptsToEmpty';
+import { navigatePromptsDown } from './flows/navigate-prompts/navigate-prompts-down';
+import { navigatePromptsUp } from './flows/navigate-prompts/navigate-prompts-up';
+import { navigatePromptsToEmpty } from './flows/navigate-prompts/navigate-prompts-to-empty';
 import { hoverOverLink } from './flows/link-hover-preview';
 import { showFileTree } from './flows/file-tree/show-file-tree';
 import { collapseExpandFileTree } from './flows/file-tree/collapse-file-tree';
